chore: deleted file with duplicate test.
