--- packages/dart/npt_flutter/test/features/profile/view/profile_view_widget_test.dart
@@ -1,478 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/annotations.dart';
-import 'package:mockito/mockito.dart';
-import 'package:npt_flutter/app.dart';
-import 'package:npt_flutter/features/favorite/favorite.dart';
-import 'package:npt_flutter/features/profile/profile.dart';
-import 'package:npt_flutter/features/profile/widgets/profile_delete_button.dart';
-import 'package:npt_flutter/features/profile_list/profile_list.dart';
-import 'package:npt_flutter/features/settings/settings.dart';
-import 'package:npt_flutter/localization/app_localizations.dart';
-import 'package:npt_flutter/util/language.dart';
-import 'package:npt_flutter/widgets/loader_bar.dart';
-import 'package:npt_flutter/widgets/spinner.dart';
-
-import 'profile_view_widget_test.mocks.dart';
-
-// Mock classes for testing
-@GenerateNiceMocks([
-  MockSpec<ProfileBloc>(),
-  MockSpec<SettingsBloc>(),
-  MockSpec<ProfilesSelectedCubit>(),
-  MockSpec<FavoriteBloc>(),
-  MockSpec<FavoriteRepository>(),
-])
-void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
-
-  group('ProfileView Widget Tests', () {
-    late MockProfileBloc mockProfileBloc;
-    late MockSettingsBloc mockSettingsBloc;
-    late MockProfilesSelectedCubit mockProfilesSelectedCubit;
-    late MockFavoriteBloc mockFavoriteBloc;
-    late MockFavoriteRepository mockFavoriteRepository;
-
-    const testUuid = 'test-uuid-123';
-    const testProfile = Profile(
-      testUuid,
-      displayName: 'Test Profile',
-      sshnpdAtsign: '@test_device',
-      deviceName: 'test-device',
-      remotePort: 22,
-      localPort: 2222,
-      relayAtsign: '@relay_test',
-    );
-
-    const testSettings = Settings(
-      relayAtsign: '@rv_eu',
-      overrideRelay: false,
-      viewLayout: PreferredViewLayout.minimal,
-      darkMode: false,
-      language: Language.english,
-    );
-
-    setUp(() {
-      mockProfileBloc = MockProfileBloc();
-      mockSettingsBloc = MockSettingsBloc();
-      mockProfilesSelectedCubit = MockProfilesSelectedCubit();
-      mockFavoriteBloc = MockFavoriteBloc();
-      mockFavoriteRepository = MockFavoriteRepository();
-
-      // Provide dummy values for the mocks
-      provideDummy<ProfileState>(const ProfileInitial(testUuid));
-      provideDummy<SettingsState>(const SettingsInitial());
-      provideDummy<ProfilesSelectedState>(const ProfilesSelectedState({}));
-      provideDummy<FavoritesState>(const FavoritesInitial());
-    });
-
-    tearDown(() {
-      reset(mockProfileBloc);
-      reset(mockSettingsBloc);
-      reset(mockProfilesSelectedCubit);
-      reset(mockFavoriteBloc);
-      reset(mockFavoriteRepository);
-    });
-
-    Widget createTestWidget({
-      required ProfileState profileState,
-      required SettingsState settingsState,
-      ProfilesSelectedState? profilesSelectedState,
-      FavoritesState? favoritesState,
-    }) {
-      // Setup the bloc state mocks directly
-      when(mockProfileBloc.state).thenReturn(profileState);
-      when(mockSettingsBloc.state).thenReturn(settingsState);
-      when(mockProfilesSelectedCubit.state).thenReturn(
-        profilesSelectedState ?? const ProfilesSelectedState({}),
-      );
-      when(mockFavoriteBloc.state).thenReturn(
-        favoritesState ?? const FavoritesInitial(),
-      );
-
-      // Setup the stream mocks
-      when(mockProfileBloc.stream).thenAnswer((_) => Stream.value(profileState));
-      when(mockSettingsBloc.stream).thenAnswer((_) => Stream.value(settingsState));
-      when(mockProfilesSelectedCubit.stream).thenAnswer(
-        (_) => Stream.value(profilesSelectedState ?? const ProfilesSelectedState({})),
-      );
-      when(mockFavoriteBloc.stream).thenAnswer(
-        (_) => Stream.value(favoritesState ?? const FavoritesInitial()),
-      );
-
-      return MaterialApp(
-        // Use the App's navigation key to provide proper context for SizeConfig
-        navigatorKey: App.navState,
-        localizationsDelegates: AppLocalizations.localizationsDelegates,
-        supportedLocales: AppLocalizations.supportedLocales,
-        home: Scaffold(
-          body: MultiBlocProvider(
-            providers: [
-              BlocProvider<ProfileBloc>.value(value: mockProfileBloc),
-              BlocProvider<SettingsBloc>.value(value: mockSettingsBloc),
-              BlocProvider<ProfilesSelectedCubit>.value(value: mockProfilesSelectedCubit),
-              BlocProvider<FavoriteBloc>.value(value: mockFavoriteBloc),
-            ],
-            child: const ProfileView(),
-          ),
-        ),
-      );
-    }
-
-    group('ProfileInitial State', () {
-      testWidgets('should trigger ProfileLoadEvent when state is ProfileInitial', (tester) async {
-        const profileState = ProfileInitial(testUuid);
-        const settingsState = SettingsLoaded(settings: testSettings);
-
-        await tester.pumpWidget(createTestWidget(
-          profileState: profileState,
-          settingsState: settingsState,
-        ));
-        await tester.pump();
-
-        verify(mockProfileBloc.add(const ProfileLoadEvent())).called(greaterThanOrEqualTo(1));
-      });
-
-      testWidgets('should display LoaderBar and ProfileRefreshButton', (tester) async {
-        const profileState = ProfileInitial(testUuid);
-        const settingsState = SettingsLoaded(settings: testSettings);
-
-        await tester.pumpWidget(createTestWidget(
-          profileState: profileState,
-          settingsState: settingsState,
-        ));
-        await tester.pump();
-
-        expect(find.byType(LoaderBar), findsOneWidget);
-        expect(find.byType(ProfileRefreshButton), findsOneWidget);
-        expect(find.byType(Row), findsWidgets);
-      });
-    });
-
-    group('ProfileLoading State', () {
-      testWidgets('should display LoaderBar and ProfileRefreshButton', (tester) async {
-        const profileState = ProfileLoading(testUuid);
-        const settingsState = SettingsLoaded(settings: testSettings);
-
-        await tester.pumpWidget(createTestWidget(
-          profileState: profileState,
-          settingsState: settingsState,
-        ));
-        await tester.pump();
-
-        expect(find.byType(LoaderBar), findsOneWidget);
-        expect(find.byType(ProfileRefreshButton), findsOneWidget);
-        expect(find.byType(Row), findsWidgets);
-      });
-
-      testWidgets('should center the loading content', (tester) async {
-        const profileState = ProfileLoading(testUuid);
-        const settingsState = SettingsLoaded(settings: testSettings);
-
-        await tester.pumpWidget(createTestWidget(
-          profileState: profileState,
-          settingsState: settingsState,
-        ));
-        await tester.pump();
-
-        final rowWidget = tester.widget<Row>(find.byType(Row));
-        expect(rowWidget.mainAxisAlignment, MainAxisAlignment.center);
-      });
-    });
-
-    group('ProfileFailedLoad State', () {
-      testWidgets('should display error message and ProfileDeleteButton', (tester) async {
-        const profileState = ProfileFailedLoad(testUuid);
-        const settingsState = SettingsLoaded(settings: testSettings);
-
-        await tester.pumpWidget(createTestWidget(
-          profileState: profileState,
-          settingsState: settingsState,
-        ));
-        await tester.pump();
-
-        // Get the context to access localized strings
-        final BuildContext context = tester.element(find.byType(ProfileView));
-        final String expectedErrorMessage = AppLocalizations.of(context)!.errorProfileLoadFailed;
-
-        // Find the exact localized error message
-        expect(find.text(expectedErrorMessage), findsOneWidget);
-
-        expect(find.byType(ProfileDeleteButton), findsOneWidget);
-        expect(find.byType(Row), findsOneWidget);
-      });
-
-      testWidgets('should space the error content', (tester) async {
-        const profileState = ProfileFailedLoad(testUuid);
-        const settingsState = SettingsLoaded(settings: testSettings);
-
-        await tester.pumpWidget(createTestWidget(
-          profileState: profileState,
-          settingsState: settingsState,
-        ));
-        await tester.pump();
-
-        final rowWidget = tester.widget<Row>(find.byType(Row));
-        expect(rowWidget.mainAxisAlignment, MainAxisAlignment.spaceBetween);
-      });
-    });
-
-    group('ProfileLoadedState with Settings', () {
-      group('Minimal View Layout', () {
-        testWidgets('should display ProfileViewMinimal when viewLayout is minimal', (tester) async {
-          const profileState = ProfileLoaded(testUuid, profile: testProfile);
-          const settingsState = SettingsLoaded(settings: testSettings);
-
-          await tester.pumpWidget(createTestWidget(
-            profileState: profileState,
-            settingsState: settingsState,
-          ));
-          await tester.pump();
-
-          expect(find.byType(ProfileViewMinimal), findsOneWidget);
-        });
-      });
-
-      group('SSH Style View Layout', () {
-        testWidgets('should display ProfileViewSshStyle when viewLayout is sshStyle', (tester) async {
-          const profileState = ProfileLoaded(testUuid, profile: testProfile);
-          const settingsWithSshStyle = Settings(
-            relayAtsign: '@rv_eu',
-            overrideRelay: false,
-            viewLayout: PreferredViewLayout.sshStyle,
-            darkMode: false,
-            language: Language.english,
-          );
-          const settingsState = SettingsLoaded(settings: settingsWithSshStyle);
-
-          await tester.pumpWidget(createTestWidget(
-            profileState: profileState,
-            settingsState: settingsState,
-          ));
-          await tester.pump();
-
-          expect(find.byType(ProfileViewSshStyle), findsOneWidget);
-        });
-      });
-
-      group('No Settings State (null viewLayout)', () {
-        testWidgets('should display Spinner when SettingsState is not SettingsLoadedState', (tester) async {
-          const profileState = ProfileLoaded(testUuid, profile: testProfile);
-          const settingsState = SettingsInitial();
-
-          await tester.pumpWidget(createTestWidget(
-            profileState: profileState,
-            settingsState: settingsState,
-          ));
-          await tester.pump();
-
-          expect(find.byType(Spinner), findsOneWidget);
-          expect(find.byType(Center), findsOneWidget);
-        });
-
-        testWidgets('should display Spinner when SettingsState is SettingsLoading', (tester) async {
-          const profileState = ProfileLoaded(testUuid, profile: testProfile);
-          const settingsState = SettingsLoading();
-
-          await tester.pumpWidget(createTestWidget(
-            profileState: profileState,
-            settingsState: settingsState,
-          ));
-          await tester.pump();
-
-          expect(find.byType(Spinner), findsOneWidget);
-          expect(find.byType(Center), findsOneWidget);
-        });
-      });
-    });
-
-    group('ProfileViewMinimal Widget Tests', () {
-      Widget createMinimalTestWidget({
-        ProfileState? profileState,
-        ProfilesSelectedState? profilesSelectedState,
-        FavoritesState? favoritesState,
-      }) {
-        // Setup the bloc state mocks
-        when(mockProfileBloc.state).thenReturn(
-          profileState ?? const ProfileLoaded(testUuid, profile: testProfile),
-        );
-        when(mockProfilesSelectedCubit.state).thenReturn(
-          profilesSelectedState ?? const ProfilesSelectedState({}),
-        );
-        when(mockFavoriteBloc.state).thenReturn(
-          favoritesState ?? const FavoritesInitial(),
-        );
-
-        // Setup the stream mocks
-        when(mockProfileBloc.stream).thenAnswer(
-          (_) => Stream.value(profileState ?? const ProfileLoaded(testUuid, profile: testProfile)),
-        );
-        when(mockProfilesSelectedCubit.stream).thenAnswer(
-          (_) => Stream.value(profilesSelectedState ?? const ProfilesSelectedState({})),
-        );
-        when(mockFavoriteBloc.stream).thenAnswer(
-          (_) => Stream.value(favoritesState ?? const FavoritesInitial()),
-        );
-
-        return MaterialApp(
-          navigatorKey: App.navState,
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          supportedLocales: AppLocalizations.supportedLocales,
-          home: Scaffold(
-            body: MultiBlocProvider(
-              providers: [
-                BlocProvider<ProfileBloc>.value(value: mockProfileBloc),
-                BlocProvider<SettingsBloc>.value(value: mockSettingsBloc),
-                BlocProvider<ProfilesSelectedCubit>.value(value: mockProfilesSelectedCubit),
-                BlocProvider<FavoriteBloc>.value(value: mockFavoriteBloc),
-              ],
-              child: const ProfileViewMinimal(),
-            ),
-          ),
-        );
-      }
-
-      testWidgets('should render without ProviderNotFoundException', (tester) async {
-        await tester.pumpWidget(createMinimalTestWidget());
-        await tester.pump();
-
-        expect(find.byType(ProfileViewMinimal), findsOneWidget);
-        expect(find.byType(Row), findsOneWidget);
-      });
-
-      testWidgets('should display all expected child widgets', (tester) async {
-        await tester.pumpWidget(createMinimalTestWidget());
-        await tester.pump();
-
-        // Check for presence of key widgets within ProfileViewMinimal
-        expect(find.byType(ProfileSelectBox), findsOneWidget);
-        expect(find.byType(ProfileDisplayName), findsOneWidget);
-        expect(find.byType(ProfileStatusIndicator), findsOneWidget);
-        expect(find.byType(ProfileRunButton), findsOneWidget);
-        expect(find.byType(ProfileFavoriteButton), findsOneWidget);
-        expect(find.byType(ProfilePopupMenuButton), findsOneWidget);
-      });
-
-      testWidgets('should handle selected profile state', (tester) async {
-        const selectedState = ProfilesSelectedState({testUuid});
-
-        await tester.pumpWidget(createMinimalTestWidget(
-          profilesSelectedState: selectedState,
-        ));
-        await tester.pump();
-
-        expect(find.byType(ProfileViewMinimal), findsOneWidget);
-        expect(find.byType(ProfileSelectBox), findsOneWidget);
-      });
-
-      testWidgets('should handle favorite profile state', (tester) async {
-        const favoriteProfile = FavoriteProfile(uuid: testUuid);
-        const favoritesState = FavoritesLoaded([favoriteProfile]);
-
-        await tester.pumpWidget(createMinimalTestWidget(
-          favoritesState: favoritesState,
-        ));
-        await tester.pump();
-
-        expect(find.byType(ProfileViewMinimal), findsOneWidget);
-        expect(find.byType(ProfileFavoriteButton), findsOneWidget);
-      });
-    });
-
-    group('ProfileViewSshStyle Widget Tests', () {
-      Widget createSshStyleTestWidget({
-        ProfileState? profileState,
-        ProfilesSelectedState? profilesSelectedState,
-        FavoritesState? favoritesState,
-      }) {
-        // Setup the bloc state mocks
-        when(mockProfileBloc.state).thenReturn(
-          profileState ?? const ProfileLoaded(testUuid, profile: testProfile),
-        );
-        when(mockProfilesSelectedCubit.state).thenReturn(
-          profilesSelectedState ?? const ProfilesSelectedState({}),
-        );
-        when(mockFavoriteBloc.state).thenReturn(
-          favoritesState ?? const FavoritesInitial(),
-        );
-
-        // Setup the stream mocks
-        when(mockProfileBloc.stream).thenAnswer(
-          (_) => Stream.value(profileState ?? const ProfileLoaded(testUuid, profile: testProfile)),
-        );
-        when(mockProfilesSelectedCubit.stream).thenAnswer(
-          (_) => Stream.value(profilesSelectedState ?? const ProfilesSelectedState({})),
-        );
-        when(mockFavoriteBloc.stream).thenAnswer(
-          (_) => Stream.value(favoritesState ?? const FavoritesInitial()),
-        );
-
-        return MaterialApp(
-          navigatorKey: App.navState,
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          supportedLocales: AppLocalizations.supportedLocales,
-          home: Scaffold(
-            body: MultiBlocProvider(
-              providers: [
-                BlocProvider<ProfileBloc>.value(value: mockProfileBloc),
-                BlocProvider<SettingsBloc>.value(value: mockSettingsBloc),
-                BlocProvider<ProfilesSelectedCubit>.value(value: mockProfilesSelectedCubit),
-                BlocProvider<FavoriteBloc>.value(value: mockFavoriteBloc),
-              ],
-              child: const ProfileViewSshStyle(),
-            ),
-          ),
-        );
-      }
-
-      testWidgets('should render without ProviderNotFoundException', (tester) async {
-        await tester.pumpWidget(createSshStyleTestWidget());
-        await tester.pump();
-
-        expect(find.byType(ProfileViewSshStyle), findsOneWidget);
-        expect(find.byType(Row), findsOneWidget);
-      });
-
-      testWidgets('should display all expected child widgets', (tester) async {
-        await tester.pumpWidget(createSshStyleTestWidget());
-        await tester.pump();
-
-        // Check for presence of key widgets within ProfileViewSshStyle
-        expect(find.byType(ProfileSelectBox), findsOneWidget);
-        expect(find.byType(ProfileDisplayName), findsOneWidget);
-        expect(find.byType(ProfileDeviceName), findsOneWidget);
-        expect(find.byType(ProfileServiceView), findsOneWidget);
-        expect(find.byType(ProfileStatusIndicator), findsOneWidget);
-        expect(find.byType(ProfileRunButton), findsOneWidget);
-        expect(find.byType(ProfileFavoriteButton), findsOneWidget);
-        expect(find.byType(ProfilePopupMenuButton), findsOneWidget);
-      });
-
-      testWidgets('should handle selected profile state', (tester) async {
-        const selectedState = ProfilesSelectedState({testUuid});
-
-        await tester.pumpWidget(createSshStyleTestWidget(
-          profilesSelectedState: selectedState,
-        ));
-        await tester.pump();
-
-        expect(find.byType(ProfileViewSshStyle), findsOneWidget);
-        expect(find.byType(ProfileSelectBox), findsOneWidget);
-      });
-
-      testWidgets('should handle favorite profile state', (tester) async {
-        const favoriteProfile = FavoriteProfile(uuid: testUuid);
-        const favoritesState = FavoritesLoaded([favoriteProfile]);
-
-        await tester.pumpWidget(createSshStyleTestWidget(
-          favoritesState: favoritesState,
-        ));
-        await tester.pump();
-
-        expect(find.byType(ProfileViewSshStyle), findsOneWidget);
-        expect(find.byType(ProfileFavoriteButton), findsOneWidget);
-      });
-    });
-  });
-}
