--- packages/dart/npt_flutter/test/features/profile/view/profile_view_widget_test.mocks.dart
@@ -1,639 +1,0 @@
-// Mocks generated by Mockito 5.4.6 from annotations
-// in npt_flutter/test/features/profile/view/profile_view_widget_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i4;
-
-import 'package:flutter_bloc/flutter_bloc.dart' as _i5;
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:mockito/src/dummies.dart' as _i3;
-import 'package:npt_flutter/app.dart' as _i9;
-import 'package:npt_flutter/features/favorite/favorite.dart' as _i8;
-import 'package:npt_flutter/features/profile/profile.dart' as _i2;
-import 'package:npt_flutter/features/profile_list/profile_list.dart' as _i7;
-import 'package:npt_flutter/features/settings/settings.dart' as _i6;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: must_be_immutable
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [ProfileBloc].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockProfileBloc extends _i1.Mock implements _i2.ProfileBloc {
-  @override
-  String get uuid => (super.noSuchMethod(
-        Invocation.getter(#uuid),
-        returnValue: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#uuid),
-        ),
-        returnValueForMissingStub: _i3.dummyValue<String>(
-          this,
-          Invocation.getter(#uuid),
-        ),
-      ) as String);
-
-  @override
-  _i2.ProfileState get state => (super.noSuchMethod(
-        Invocation.getter(#state),
-        returnValue: _i3.dummyValue<_i2.ProfileState>(
-          this,
-          Invocation.getter(#state),
-        ),
-        returnValueForMissingStub: _i3.dummyValue<_i2.ProfileState>(
-          this,
-          Invocation.getter(#state),
-        ),
-      ) as _i2.ProfileState);
-
-  @override
-  _i4.Stream<_i2.ProfileState> get stream => (super.noSuchMethod(
-        Invocation.getter(#stream),
-        returnValue: _i4.Stream<_i2.ProfileState>.empty(),
-        returnValueForMissingStub: _i4.Stream<_i2.ProfileState>.empty(),
-      ) as _i4.Stream<_i2.ProfileState>);
-
-  @override
-  bool get isClosed => (super.noSuchMethod(
-        Invocation.getter(#isClosed),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-
-  @override
-  void onEvent(_i2.ProfileEvent? event) => super.noSuchMethod(
-        Invocation.method(
-          #onEvent,
-          [event],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void emit(_i2.ProfileState? state) => super.noSuchMethod(
-        Invocation.method(
-          #emit,
-          [state],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void add(_i2.ProfileEvent? event) => super.noSuchMethod(
-        Invocation.method(
-          #add,
-          [event],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void on<E extends _i2.ProfileEvent>(
-    _i5.EventHandler<E, _i2.ProfileState>? handler, {
-    _i5.EventTransformer<E>? transformer,
-  }) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #on,
-          [handler],
-          {#transformer: transformer},
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onTransition(
-          _i5.Transition<_i2.ProfileEvent, _i2.ProfileState>? transition) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onTransition,
-          [transition],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  _i4.Future<void> close() => (super.noSuchMethod(
-        Invocation.method(
-          #close,
-          [],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-
-  @override
-  void onChange(_i5.Change<_i2.ProfileState>? change) => super.noSuchMethod(
-        Invocation.method(
-          #onChange,
-          [change],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void addError(
-    Object? error, [
-    StackTrace? stackTrace,
-  ]) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #addError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onError(
-    Object? error,
-    StackTrace? stackTrace,
-  ) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-}
-
-/// A class which mocks [SettingsBloc].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockSettingsBloc extends _i1.Mock implements _i6.SettingsBloc {
-  @override
-  _i6.SettingsState get state => (super.noSuchMethod(
-        Invocation.getter(#state),
-        returnValue: _i3.dummyValue<_i6.SettingsState>(
-          this,
-          Invocation.getter(#state),
-        ),
-        returnValueForMissingStub: _i3.dummyValue<_i6.SettingsState>(
-          this,
-          Invocation.getter(#state),
-        ),
-      ) as _i6.SettingsState);
-
-  @override
-  _i4.Stream<_i6.SettingsState> get stream => (super.noSuchMethod(
-        Invocation.getter(#stream),
-        returnValue: _i4.Stream<_i6.SettingsState>.empty(),
-        returnValueForMissingStub: _i4.Stream<_i6.SettingsState>.empty(),
-      ) as _i4.Stream<_i6.SettingsState>);
-
-  @override
-  bool get isClosed => (super.noSuchMethod(
-        Invocation.getter(#isClosed),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-
-  @override
-  void clear() => super.noSuchMethod(
-        Invocation.method(
-          #clear,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onEvent(_i6.SettingsEvent? event) => super.noSuchMethod(
-        Invocation.method(
-          #onEvent,
-          [event],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void emit(_i6.SettingsState? state) => super.noSuchMethod(
-        Invocation.method(
-          #emit,
-          [state],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void add(_i6.SettingsEvent? event) => super.noSuchMethod(
-        Invocation.method(
-          #add,
-          [event],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void on<E extends _i6.SettingsEvent>(
-    _i5.EventHandler<E, _i6.SettingsState>? handler, {
-    _i5.EventTransformer<E>? transformer,
-  }) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #on,
-          [handler],
-          {#transformer: transformer},
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onTransition(
-          _i5.Transition<_i6.SettingsEvent, _i6.SettingsState>? transition) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onTransition,
-          [transition],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  _i4.Future<void> close() => (super.noSuchMethod(
-        Invocation.method(
-          #close,
-          [],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-
-  @override
-  void onChange(_i5.Change<_i6.SettingsState>? change) => super.noSuchMethod(
-        Invocation.method(
-          #onChange,
-          [change],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void addError(
-    Object? error, [
-    StackTrace? stackTrace,
-  ]) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #addError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onError(
-    Object? error,
-    StackTrace? stackTrace,
-  ) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-}
-
-/// A class which mocks [ProfilesSelectedCubit].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockProfilesSelectedCubit extends _i1.Mock
-    implements _i7.ProfilesSelectedCubit {
-  @override
-  _i7.ProfilesSelectedState get state => (super.noSuchMethod(
-        Invocation.getter(#state),
-        returnValue: _i3.dummyValue<_i7.ProfilesSelectedState>(
-          this,
-          Invocation.getter(#state),
-        ),
-        returnValueForMissingStub: _i3.dummyValue<_i7.ProfilesSelectedState>(
-          this,
-          Invocation.getter(#state),
-        ),
-      ) as _i7.ProfilesSelectedState);
-
-  @override
-  _i4.Stream<_i7.ProfilesSelectedState> get stream => (super.noSuchMethod(
-        Invocation.getter(#stream),
-        returnValue: _i4.Stream<_i7.ProfilesSelectedState>.empty(),
-        returnValueForMissingStub:
-            _i4.Stream<_i7.ProfilesSelectedState>.empty(),
-      ) as _i4.Stream<_i7.ProfilesSelectedState>);
-
-  @override
-  bool get isClosed => (super.noSuchMethod(
-        Invocation.getter(#isClosed),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-
-  @override
-  void select(String? uuid) => super.noSuchMethod(
-        Invocation.method(
-          #select,
-          [uuid],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void deselect(String? uuid) => super.noSuchMethod(
-        Invocation.method(
-          #deselect,
-          [uuid],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void deselectAll() => super.noSuchMethod(
-        Invocation.method(
-          #deselectAll,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void selectAll() => super.noSuchMethod(
-        Invocation.method(
-          #selectAll,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void emit(_i7.ProfilesSelectedState? state) => super.noSuchMethod(
-        Invocation.method(
-          #emit,
-          [state],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onChange(_i5.Change<_i7.ProfilesSelectedState>? change) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onChange,
-          [change],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void addError(
-    Object? error, [
-    StackTrace? stackTrace,
-  ]) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #addError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onError(
-    Object? error,
-    StackTrace? stackTrace,
-  ) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  _i4.Future<void> close() => (super.noSuchMethod(
-        Invocation.method(
-          #close,
-          [],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-}
-
-/// A class which mocks [FavoriteBloc].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockFavoriteBloc extends _i1.Mock implements _i8.FavoriteBloc {
-  @override
-  _i8.FavoritesState get state => (super.noSuchMethod(
-        Invocation.getter(#state),
-        returnValue: _i3.dummyValue<_i8.FavoritesState>(
-          this,
-          Invocation.getter(#state),
-        ),
-        returnValueForMissingStub: _i3.dummyValue<_i8.FavoritesState>(
-          this,
-          Invocation.getter(#state),
-        ),
-      ) as _i8.FavoritesState);
-
-  @override
-  _i4.Stream<_i8.FavoritesState> get stream => (super.noSuchMethod(
-        Invocation.getter(#stream),
-        returnValue: _i4.Stream<_i8.FavoritesState>.empty(),
-        returnValueForMissingStub: _i4.Stream<_i8.FavoritesState>.empty(),
-      ) as _i4.Stream<_i8.FavoritesState>);
-
-  @override
-  bool get isClosed => (super.noSuchMethod(
-        Invocation.getter(#isClosed),
-        returnValue: false,
-        returnValueForMissingStub: false,
-      ) as bool);
-
-  @override
-  void clearAll() => super.noSuchMethod(
-        Invocation.method(
-          #clearAll,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onEvent(_i8.FavoriteEvent? event) => super.noSuchMethod(
-        Invocation.method(
-          #onEvent,
-          [event],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void emit(_i8.FavoritesState? state) => super.noSuchMethod(
-        Invocation.method(
-          #emit,
-          [state],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void add(_i8.FavoriteEvent? event) => super.noSuchMethod(
-        Invocation.method(
-          #add,
-          [event],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void on<E extends _i8.FavoriteEvent>(
-    _i5.EventHandler<E, _i8.FavoritesState>? handler, {
-    _i5.EventTransformer<E>? transformer,
-  }) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #on,
-          [handler],
-          {#transformer: transformer},
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onTransition(
-          _i5.Transition<_i8.FavoriteEvent, _i8.FavoritesState>? transition) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onTransition,
-          [transition],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  _i4.Future<void> close() => (super.noSuchMethod(
-        Invocation.method(
-          #close,
-          [],
-        ),
-        returnValue: _i4.Future<void>.value(),
-        returnValueForMissingStub: _i4.Future<void>.value(),
-      ) as _i4.Future<void>);
-
-  @override
-  void onChange(_i5.Change<_i8.FavoritesState>? change) => super.noSuchMethod(
-        Invocation.method(
-          #onChange,
-          [change],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void addError(
-    Object? error, [
-    StackTrace? stackTrace,
-  ]) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #addError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void onError(
-    Object? error,
-    StackTrace? stackTrace,
-  ) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #onError,
-          [
-            error,
-            stackTrace,
-          ],
-        ),
-        returnValueForMissingStub: null,
-      );
-}
-
-/// A class which mocks [FavoriteRepository].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockFavoriteRepository extends _i1.Mock
-    implements _i8.FavoriteRepository {
-  @override
-  _i4.Future<Map<String, _i8.Favorite<_i9.Loggable>>?> getFavorites(
-          {bool? useCache = true}) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #getFavorites,
-          [],
-          {#useCache: useCache},
-        ),
-        returnValue:
-            _i4.Future<Map<String, _i8.Favorite<_i9.Loggable>>?>.value(),
-        returnValueForMissingStub:
-            _i4.Future<Map<String, _i8.Favorite<_i9.Loggable>>?>.value(),
-      ) as _i4.Future<Map<String, _i8.Favorite<_i9.Loggable>>?>);
-
-  @override
-  _i4.Future<bool> addFavorite(_i8.Favorite<_i9.Loggable>? favorite) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #addFavorite,
-          [favorite],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
-
-  @override
-  _i4.Future<bool> removeFavorites(Iterable<String>? uuids) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #removeFavorites,
-          [uuids],
-        ),
-        returnValue: _i4.Future<bool>.value(false),
-        returnValueForMissingStub: _i4.Future<bool>.value(false),
-      ) as _i4.Future<bool>);
-}
